Extracted Math, Angles, and Shapes from Target (prelude to Target Extensions)
