+++ WrathCombo/CustomCombo/Functions/Target.cs
@@ -2,6 +2,7 @@
 using ECommons;
 using ECommons.DalamudServices;
 using ECommons.GameFunctions;
+using ECommons.MathHelpers;
 using ECommons.Throttlers;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using FFXIVClientStructs.FFXIV.Client.System.Framework;
@@ -15,7 +16,9 @@ using WrathCombo.Core;
 using WrathCombo.Data;
 using WrathCombo.Extensions;
 using WrathCombo.Services;
+using static WrathCombo.CustomComboNS.Functions.WrathMath;
 using ObjectKind = Dalamud.Game.ClientState.Objects.Enums.ObjectKind;
+
 namespace WrathCombo.CustomComboNS.Functions;
 
 internal abstract partial class CustomComboFunctions
@@ -39,8 +42,8 @@ internal abstract partial class CustomComboFunctions
     }
 
     /// <summary> Gets the current target or null. </summary>
-    public static IBattleChara? CurrentTarget => 
-        OverrideTarget as IBattleChara ?? 
+    public static IBattleChara? CurrentTarget =>
+        OverrideTarget as IBattleChara ??
         SimpleTarget.HardTarget;
 
     #region Target Checks
@@ -294,7 +297,7 @@ internal abstract partial class CustomComboFunctions
     public static uint GetTargetCurrentHP(IGameObject? optionalTarget = null, bool forceUsePending = false)
     {
         optionalTarget ??= CurrentTarget;
-        if (optionalTarget is IBattleChara chara) 
+        if (optionalTarget is IBattleChara chara)
         {
             if ((Service.Configuration.UseExperimentalHP || forceUsePending) && SimpleTargetState.TargetStates.TryGetFirst(x => x.GameObjectID == chara.GameObjectId, out var p))
                 return p.CurrentHP;
@@ -692,14 +695,6 @@ internal abstract partial class CustomComboFunctions
 
     #region Positional Helpers
 
-    public enum AttackAngle
-    {
-        Front,
-        Flank,
-        Rear,
-        Unknown,
-    }
-
     /// <summary> Gets the player's position relative to the target. </summary>
     /// <returns> Front, Flank, Rear or Unknown as AttackAngle type. </returns>
     public static AttackAngle AngleToTarget(IGameObject? optionalTarget = null)
@@ -712,8 +707,8 @@ internal abstract partial class CustomComboFunctions
         if (target is null)
             return AttackAngle.Unknown;
 
-        float rotation = PositionalMath.GetRotation(target.Position, player.Position) - target.Rotation;
-        float regionDegrees = PositionalMath.ToDegrees(rotation) + (rotation < 0f ? 360f : 0f);
+        Angle rotation = new(GetRotation(target.Position, player.Position) - target.Rotation);
+        float regionDegrees = rotation.Deg;
 
         return regionDegrees switch
         {
@@ -723,22 +718,6 @@ internal abstract partial class CustomComboFunctions
             >= 225f and <= 315f => AttackAngle.Flank,   // 270° ± 45°
             _ => AttackAngle.Unknown
         };
-    }
-
-    /// <summary> Performs positional calculations. Based on the excellent Resonant plugin. </summary>
-    internal static class PositionalMath
-    {
-        public const float DegToRad = MathF.PI / 180f;
-        public const float RadToDeg = 180f / MathF.PI;
-
-        public static float ToRadians(float degrees) => degrees * DegToRad;
-        public static float ToDegrees(float radians) => radians * RadToDeg;
-
-        public static float GetRotation(Vector3 a, Vector3 b) => MathF.Atan2(b.X - a.X, b.Z - a.Z);
-        public static Vector3 GetDirection(Vector3 a, Vector3 b) => ToDirection(GetRotation(a, b));
-
-        public static float ToRotation(Vector3 direction) => MathF.Atan2(direction.X, direction.Z);
-        public static Vector3 ToDirection(float rotation) => new(MathF.Sin(rotation), 0f, MathF.Cos(rotation));
     }
 
     #endregion
@@ -838,7 +817,7 @@ internal abstract partial class CustomComboFunctions
             return targets.Where(o =>
                 GetTargetDistance(o) <= size &&
                 PointInCone(o.Position - player.Position,
-                    PositionalMath.GetDirection(player.Position, target.Position),
+                    GetDirection(player.Position, target.Position),
                     45f));
 
         // Line AoEs
@@ -846,7 +825,7 @@ internal abstract partial class CustomComboFunctions
             return targets.Where(o =>
                 GetTargetDistance(o) <= size &&
                 HitboxInRect(o,
-                    PositionalMath.GetRotation(player.Position, target.Position),
+                    GetRotation(player.Position, target.Position),
                     size * 0.5f, width * 0.5f));
 
         // If it was not a supported type
@@ -902,7 +881,7 @@ internal abstract partial class CustomComboFunctions
                  o.GameObjectId == target.GameObjectId &&
                  GetTargetDistance(o) <= size &&
                  PointInCone(o.Position - LocalPlayer.Position,
-                     PositionalMath.GetDirection(LocalPlayer.Position, target.Position),
+                     WrathMath.GetDirection(LocalPlayer.Position, target.Position),
                      45f));
     }
 
@@ -918,95 +897,9 @@ internal abstract partial class CustomComboFunctions
                 o.GameObjectId == target.GameObjectId &&
                 GetTargetDistance(o) <= size &&
                 HitboxInRect(o,
-                    PositionalMath.GetRotation(LocalPlayer.Position, target.Position),
+                    GetRotation(LocalPlayer.Position, target.Position),
                     size * 0.5f, width * 0.5f));
     }
-
-    #region Shape Helpers
-
-    public interface IAoeShape { }
-    public struct SelfCircle : IAoeShape { }
-    public struct Circle : IAoeShape { }
-    public struct Cone : IAoeShape { }
-    public struct Line : IAoeShape { }
-
-    #region Point in Circle
-    public static bool PointInCircle(Vector3 offsetFromOrigin, float radius)
-    {
-        return offsetFromOrigin.LengthSquared() <= radius * radius;
-    }
-    #endregion
-
-    #region Point in Cone
-    public static bool PointInCone(Vector3 offsetFromOrigin, Vector3 direction, float halfAngle)
-    {
-        return Vector3.Dot(Vector3.Normalize(offsetFromOrigin), direction) > MathF.Cos(halfAngle);
-    }
-    #endregion
-
-    #region Point in Rect
-    public static bool HitboxInRect(IGameObject o, float rotation, float halfLength, float halfWidth)
-    {
-        if (LocalPlayer is not { } player) return false;
-
-        Vector2 A = new(player.Position.X, player.Position.Z);
-        Vector2 d = new(MathF.Sin(rotation), MathF.Cos(rotation));
-        Vector2 n = new(d.Y, -d.X);
-        Vector2 P = new(o.Position.X, o.Position.Z);
-        float R = o.HitboxRadius;
-
-        Vector2 Q = A + d * halfLength;
-        Vector2 P2 = P - Q;
-        Vector2 Ptrans = new(Vector2.Dot(P2, n), Vector2.Dot(P2, d));
-        Vector2 Pabs = new(Math.Abs(Ptrans.X), Math.Abs(Ptrans.Y));
-        Vector2 Pcorner = new(Math.Abs(Ptrans.X) - halfWidth, Math.Abs(Ptrans.Y) - halfLength);
-#if DEBUG
-        if (Svc.GameGui.WorldToScreen(o.Position, out var screenCoords))
-        {
-            var objectText = $"A = {A}\n" +
-                             $"d = {d}\n" +
-                             $"n = {n}\n" +
-                             $"P = {P}\n" +
-                             $"Q = {Q}\n" +
-                             $"P2 = {P2}\n" +
-                             $"Ptrans = {Ptrans}\n" +
-                             $"Pcorner{Pcorner}\n" +
-                             $"R = {R}, R * R = {R * R}\n" +
-                             $"PcornerSquared = {Pcorner.LengthSquared()}\n" +
-                             $"PcornerX > R = {Pcorner.X > R}, PcornerY > R = {Pcorner.Y > R}\n" +
-                             $"PcornerX <= 0 = {Pcorner.X <= 0}, PcornerY <= 0 = {Pcorner.Y <= 0}";
-
-            var screenPos = ImGui.GetMainViewport().Pos;
-
-            ImGui.SetNextWindowPos(new Vector2(screenCoords.X, screenCoords.Y));
-
-            ImGui.SetNextWindowBgAlpha(1f);
-            if (ImGui.Begin(
-                    $"Actor###ActorWindow{o.GameObjectId}",
-                    ImGuiWindowFlags.NoDecoration |
-                    ImGuiWindowFlags.AlwaysAutoResize |
-                    ImGuiWindowFlags.NoSavedSettings |
-                    ImGuiWindowFlags.NoMove |
-                    ImGuiWindowFlags.NoMouseInputs |
-                    ImGuiWindowFlags.NoDocking |
-                    ImGuiWindowFlags.NoFocusOnAppearing |
-                    ImGuiWindowFlags.NoNav))
-                ImGui.Text(objectText);
-            ImGui.End();
-        }
-#endif
-
-        if (Pcorner.X > R || Pcorner.Y > R)
-            return false;
-
-        if (Pcorner.X <= 0 || Pcorner.Y <= 0)
-            return true;
-
-        return Pcorner.LengthSquared() <= R * R;
-    }
-    #endregion
-
-    #endregion
 
     #endregion
 }
